server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Dechogt",email = "dechardinlepro@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "celery (>=5.5.2,<6.0.0)",
    "django-prometheus (>=2.3.1,<3.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "cryptography (>=45.0.3,<46.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
